% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_return_detail.R
\name{return_detail_stats}
\alias{return_detail_stats}
\title{Scrape Detailed Return Stats}
\usage{
return_detail_stats(url, player)
}
\arguments{
\item{url}{Match url}

\item{player}{Two-letter initial of player}
}
\description{
Scrape Detailed Return Stats
}
