% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_return.R
\name{return_stats}
\alias{return_stats}
\title{Scrape Return Stats}
\usage{
return_stats(url, player)
}
\arguments{
\item{url}{Match url}

\item{player}{Two-letter initial of player}
}
\description{
Scrape Return Stats
}
