% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_serve_detail.R
\name{serve_detail_stats}
\alias{serve_detail_stats}
\title{Scrape Detailed Serve Stats}
\usage{
serve_detail_stats(url, player, level)
}
\arguments{
\item{url}{Match url}

\item{player}{Two-letter initial of player}

\item{level}{One of 'Total', '1st', or '2nd'}
}
\description{
Scrape Detailed Serve Stats
}
