% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_serve.R
\name{serve_stats}
\alias{serve_stats}
\title{Scrape Serve Stats}
\usage{
serve_stats(url, player, level)
}
\arguments{
\item{url}{Match url}

\item{player}{Two-letter initial of player}

\item{level}{One of 'Total', '1st', or '2nd'}
}
\description{
Scrape Serve Stats
}
